add tox, move to std layout, move to pyproject.toml, fix mypy
