--- setup.cfg
@@ -1,35 +1,0 @@
-[metadata]
-name = aitl-cli
-version = 0.1.0
-author = Gauthier Jolly
-author_email = gauthier.jolly@canonical.com
-description = Basic Azure Image Testing for Linux API client
-long_description = file: README.md
-long_description_content_type = text/markdown
-keywords = aitl, azure
-license = GPLv3
-classifiers =
-    License :: OSI Approved :: GNU General Public License v3 (GPLv3)
-    Programming Language :: Python :: 3
-
-[options]
-install_requires =
-    requests
-    click
-package_dir=
-    =cli
-
-[options.extras_require]
-dev =
-    black
-    flake8
-
-[options.packages.find]
-where=cli
-
-[options.entry_points]
-console_scripts =
-    aitl-cli = main:cli
-
-[flake8]
-extend-ignore = E203
