+++ setup.py
@@ -1,0 +1,5 @@
+"""Legacy Python packaging entry-point."""
+
+from setuptools import setup
+
+setup()
