+++ tox.ini
@@ -1,0 +1,23 @@
+[tox]
+requires =
+    tox>=4
+envlist = py3, lint, type
+isolated_build = True
+
+[testenv:lint]
+deps =
+    .[dev]
+commands =
+    isort --check --diff src
+    black --check --diff src
+    mypy src
+
+# Non-CI environments
+[testenv:format]
+description = Run formatting tools on local files.
+deps =
+    .[dev]
+skip_install = true
+commands =
+    isort src
+    black src
